common: enable nudging and overriding
This adds flags to allow a GCS to nudge or override while pointing to a
location or tracking. This is basically a way to deconflict commands
coming from a mission against manual controls from a ground station.
Without the nudging or overriding flags set, the manual control is to
be ignored.